# Batch writing piece indexes to file db for sequential piece writing