Here are some benchmarks for that:

Benchmark command: `./farmer bench --plot-size 100G --write-pieces-size 30G`

Before:
```
Finished benchmarking.

28.44G allocated for farming
27.93G actual space pledged (which is 98.21%)
520.80M of overhead (which is 1.79%)
304.60s plotting time
98.49M/s average plotting throughput
```
After:
```
Finished benchmarking.

28.44G allocated for farming
27.93G actual space pledged (which is 98.21%)
520.70M of overhead (which is 1.79%)
293.62s plotting time
102.17M/s average plotting throughput
```

It doesn't look that important from this benchmark but with plot increasing gap between those also increase and it can become around 10-15M/s